refactor: abstracted the menu item from the navbar a bit more
