+++ src/app/pages/committees/committees.component.html
@@ -5,7 +5,7 @@
       Want to get more involved in the CSSS community? Join a committee! Committees are extensions
       of the CSSS that focuses on a specific area of interest. They are a great way to meet new
       people, gain experience, and contribute to the CSSS in a way that aligns with your interests!
-    </p> 
+    </p>
   </header>
   <section>
     <h2>🎉 Events Committee</h2>
@@ -22,10 +22,10 @@
     <h2>🌐 W3 Committee</h2>
     <p>
       The W3 committee meets asynchronously to work on the various websites and projects for the
-      CSSS such as the current website you're looking at, events pages, games, such as PacMacro, and many more. 
-      This committee is chaired by both the current System Administrator (SysAdmin) and the current
-      Webmaster of the CSSS, and all discussion currently takes place on Discord. Send the current
-      SysAdmin or Webmaster a message on Discord or reach out to them via email at
+      CSSS such as the current website you're looking at, events pages, games, such as PacMacro, and
+      many more. This committee is chaired by both the current System Administrator (SysAdmin) and
+      the current Webmaster of the CSSS, and all discussion currently takes place on Discord. Send
+      the current SysAdmin or Webmaster a message on Discord or reach out to them via email at
       <a href="mailto:csss-sysadmin@sfu.ca">csss-sysadmin&commat;sfu.ca</a> or
       <a href="csss-webmaster@sfu.ca">csss-webmaster&commat;sfu.ca</a> to join the W3 Committee!
     </p>
